refactor(security): centralize privacy boundary owners
